% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{anti_trafo}
\alias{anti_trafo}
\title{Horvath age transformation (inverse)}
\usage{
anti_trafo(x, adult_age = 20)
}
\arguments{
\item{x}{Numeric vector of raw predictions.}

\item{adult_age}{Reference adult age used by the transformation. Default
\code{20}.}
}
\value{
Numeric vector of ages in years.
}
\description{
Several clocks are trained on an age variable that is compressed on a log
scale below a reference adult age. This maps a raw prediction back to years.
}
\examples{
anti_trafo(c(-0.5, 0, 0.5, 1.5))
}
