% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{BMIQcalibration}
\alias{BMIQcalibration}
\title{BMIQ normallization}
\usage{
BMIQcalibration(
  datM,
  goldstandard.beta,
  nL = 3,
  doH = TRUE,
  nfit = 20000,
  th1.v = c(0.2, 0.75),
  th2.v = NULL,
  niter = 5,
  tol = 0.001,
  plots = FALSE,
  calibrateUnitInterval = TRUE
)
}
\arguments{
\item{datM}{data matrix}

\item{goldstandard.beta}{'goldstandard.beta'}

\item{nL}{default value is 3}

\item{doH}{default value is TRUE}

\item{nfit}{devault value id 20000}

\item{th1.v}{default value is c(0.2, 0.75)}

\item{th2.v}{default valued is NULL}

\item{niter}{default value is 5}

\item{tol}{defalt value is 0.001}

\item{plot}{default value is FALE}

\item{calibrateUnitInverval}{default value is TRuE}
}
\value{
the normalized table and a plot when plot is equal to TRUE
}
\description{
BMIQ normallization
}
\details{
ORIGINAL AUTHOR: Andrew Teschendorff
The original BMIQ function from Teschendorff 2013 adjusts for the
type-2 bias in Illumina Infinium 450k data.
Later functions and edits were provided by yours truly, Steve Horvath.
I changed the code so that one can calibrate methylation data to a gold standard.
Specifically, I took version v_1.2 by Teschendorff  and fixed minor issues.
Also I made the code more robust e.g. by changing the optimization algorithm.
Toward this end, I used the method="Nelder-Mead" in optim()

Later functions and edits by Steve Horvath
Steve Horvath took version v_1.2 by Teschendorff
and fixed minor errors. Also he made the code more robust.
Importantly, SH changed the optimization algorithm to make it more robust.
SH used method="Nelder-Mead" in optim() since the other optimization method
sometimes gets stuck. Toward this end, the function blc was replaced by blc2.
}
