% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_registry.R
\name{clock_registry}
\alias{clock_registry}
\title{The clock registry}
\usage{
clock_registry()
}
\value{
A named list of registry entries (objects of class
  \code{clock_entry}), ordered by name.
}
\description{
The clock registry
}
\examples{
reg <- clock_registry()
length(reg)
names(reg)[seq_len(5)]
reg[["Horvath"]]
}
\seealso{
\code{\link{clock_register}}, \code{\link{clock_list}}
}
