% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_registry.R
\docType{data}
\name{clock_targets}
\alias{clock_targets}
\title{Supported biological targets}
\format{
An object of class \code{character} of length 10.
}
\usage{
clock_targets
}
\value{
A character vector of the supported target names.
}
\description{
The quantity a clock estimates. A clock's \code{target} must be one of these.
}
\examples{
clock_targets
}
\keyword{datasets}
