% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefBlup}
\alias{coefBlup}
\title{BLUP clock coefficients}
\format{
A data frame with 392 CpGs and 2 variables
}
\usage{
data(coefBlup)
}
\description{
A dataset containing the coefficients and CpGs information from  Best Linear
Unbiased Prediction (BLUP) clock
}
\examples{
data(coefBlup)
}
\references{
\url{https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-019-0667-1}
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient }
}
}
\keyword{datasets}
