% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefEN}
\alias{coefEN}
\title{EN clock coefficients}
\format{
A data frame with 392 CpGs and 2 variables
}
\usage{
data(coefEN)
}
\description{
A dataset containing the coefficients and CpGs information from Elastic Net
(EN) clock
}
\examples{
data(coefEN)
}
\references{
\url{https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-019-0667-1}
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient }
}
}
\keyword{datasets}
