% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefEPIC}
\alias{coefEPIC}
\title{EPIC Gestational Age clock coefficients}
\format{
A data frame with 176 CpGs and 3 variables
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient}
  \item{UCSC_RefGene_Name}{UCSC RefGene Name}
}
}
\usage{
data(coefEPIC)
}
\description{
A dataset containing the coefficients and CpGs information from EPIC GA clock
}
\examples{
data(coefEPIC)
}
\references{
\url{https://doi.org/10.1186/s13148-021-01055-z}
}
\keyword{datasets}
