% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefHannum}
\alias{coefHannum}
\title{Hannum clock coefficients}
\format{
A data frame with 71 rows and 6 variables
}
\usage{
data(coefHannum)
}
\description{
A dataset containing the coefficients and CpGs information from Hannum clock
}
\examples{
data(coefHannum)
}
\references{
\url{https://doi.org/10.1016/j.molcel.2012.10.016}
\describe{
  \item{CpGMarker}{CpGs id}
  \item{Chrom}{Chromosome}
  \item{Pos}{Position}
  \item{Genes}{Genes}
  \item{CpGIsland }{Is CpG Island}
  \item{CoefficientTraining}{Coefficient from training }
}
}
\keyword{datasets}
