% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefKnightGA}
\alias{coefKnightGA}
\title{Knight Gestational Age clock coefficients}
\format{
A data frame with 149 CpGs and 2 variables
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient}
}
}
\usage{
data(coefKnightGA)
}
\description{
A dataset containing the coefficients and CpGs information from Knight GA clock
}
\examples{
data(coefKnightGA)
}
\references{
\url{https://dx.doi.org/10.1186\%2Fs13059-016-1068-z}
}
\keyword{datasets}
