% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefLeeGA}
\alias{coefLeeGA}
\title{Lee Gestational Age clock coefficients}
\format{
A data frame with 1126 CpGs and 5 variables
\describe{
  \item{CpGMarker}{CpGs id}
  \item{CoefficientTraining}{RPC Coefficient }
  \item{Coefficient_CPC}{}
  \item{Coefficient_refined_RPC}{ }
  \item{Coefficient_sex_classifier}{}
}
}
\usage{
data(coefLeeGA)
}
\description{
A dataset containing the coefficients and CpGs information from Lee GA clock
}
\references{
\url{https://dx.doi.org/10.18632\%2Faging.102049}
}
\keyword{datasets}
