% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefLevine}
\alias{coefLevine}
\title{Levine clock coefficients}
\format{
A data frame with 514 CpGs and 9 variables
}
\usage{
data(coefLevine)
}
\description{
A dataset containing the coefficients and CpGs information from Levine clock
}
\examples{
data(coefLevine)
}
\references{
\url{https://dx.doi.org/10.18632\%2Faging.101414}
\describe{
  \item{CpGMarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient}
  \item{Chromosome}{Chromosome}
  \item{Map.Info}{Map.Info}
  \item{Gene.Symbol}{Gene symbol}
  \item{Entrez.ID}{Entrez ID}
  \item{Univariate.Age.Correlation}{}
  \item{Horvath.Overlap}{Horvath.Overlap}
  \item{Hannum.Overlap}{Hannum.Overlap}
}
}
\keyword{datasets}
