% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefNEOaPMA450K}
\alias{coefNEOaPMA450K}
\title{NEOage PostMenstrual age Clock for 450K - Gestational Age clock coefficients}
\format{
A data frame with 409 CpGs and 2 variables
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient}
}
}
\usage{
data(coefNEOaPMA450K)
}
\description{
A dataset containing the coefficients and CpGs information from NEOage 
PostMenstrual age Clock for 450K array data.
}
\examples{
data(coefNEOaPMA450K)
}
\references{
\url{https://doi.org/10.18632/aging.203637}
}
\keyword{datasets}
