% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefNEOaPMAEPIC}
\alias{coefNEOaPMAEPIC}
\title{NEOage PostMenstrual age Clock for EPIC - Gestational Age clock coefficients}
\format{
A data frame with 522 CpGs and 2 variables
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient}
}
}
\usage{
data(coefNEOaPMAEPIC)
}
\description{
A dataset containing the coefficients and CpGs information from NEOage 
PostMenstrual age Clock for EPIC array data.
}
\examples{
data(coefNEOaPMAEPIC)
}
\references{
\url{https://doi.org/10.18632/aging.203637}
}
\keyword{datasets}
