% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefPedBE}
\alias{coefPedBE}
\title{PedBE clock coefficients}
\format{
A data frame with 1126 CpGs and 5 variables
\describe{
  \item{CpGmarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient }
  \item{corAgeTraining}{Correlation age training}
  \item{corAgeTest}{Correlation age test}
}
}
\usage{
data(coefPedBE)
}
\description{
A dataset containing the coefficients and CpGs information from PedBE clock
}
\examples{
data(coefPedBE)
}
\references{
\url{https://doi.org/10.1073/pnas.1820843116}
}
\keyword{datasets}
