% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefSkin}
\alias{coefSkin}
\title{Skin clock coefficients}
\format{
A data frame with 392 CpGs and 2 variables
\describe{
  \item{CpGMarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient }
}
}
\usage{
data(coefSkin)
}
\description{
A dataset containing the coefficients and CpGs information from Skin clock
}
\keyword{datasets}
