% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefTL}
\alias{coefTL}
\title{Telomere Length clock coefficients}
\format{
A data frame with 141 CpGs and 22 variables
}
\usage{
data(coefTL)
}
\description{
A dataset containing the coefficients and CpGs information from Telomere Length clock
}
\references{
\url{https://dx.doi.org/10.18632\%2Faging.102173}
\describe{
  \item{CpGMarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient}
  \item{CHR}{Chromosome}
  \item{bp19}{}
  \item{SourceSeq}{}
  \item{Strand}{}
  \item{Probe.SNP.based.on.Illumina.annotation}{}
  \item{Gene}{}
  \item{location}{}
  \item{Relation.to..CpGIsland}{}
  \item{Probe.SNP.based.on.1000.Genome}{}
  \item{SNP}{}
  \item{CHR.SNP}{}
  \item{bp.SNP}{}
  \item{ALLELE}{}
  \item{BSGS.P}{}
  \item{BSGS.EFFECT}{}
  \item{BSGS.H2}{}
  \item{LBC.P}{}
  \item{LBC.EFFECT}{}
  \item{LBC.R2}{}
  \item{mQTL}{}
}
}
\keyword{datasets}
