% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonClockCpgs.R
\name{commonClockCpgs}
\alias{commonClockCpgs}
\title{Get common CpGs}
\usage{
commonClockCpgs(object, clock)
}
\arguments{
\item{object}{resulting object from checkClocks functions}

\item{clock}{string with the implemented clock, possible values are : 
"Knight", "Bohlin", "Mayne" and "Lee", "Horvath", "Hannum", "Levine",
 "Horvath2", "PedBE", "Wu" and "TL"}
}
\value{
The common CpGs between input data and defined GA clock
}
\description{
Show the required CpGs contained on input data for the implemented clocks
}
\examples{
TestDataset <- get_TestDataset()
cpgs.missing.GA <- checkClocksGA(TestDataset)
cpgs.missing <- checkClocks(TestDataset)
commonClockCpgs(cpgs.missing.GA, "Bohlin")
commonClockCpgs(cpgs.missing, "Hannum")
}
