% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpgs.bn}
\alias{cpgs.bn}
\title{BN clock CpGs}
\format{
A data frame with 354 CpGs and 23 variables :
\describe{
  \item{cpgs.bn}{CpGs id}
}
}
\usage{
data(cpgs.bn)
}
\description{
An array containing the CpGs list from BN clock
}
\examples{
data(cpgs.bn)
}
\keyword{datasets}
