% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil.list.cell.type.references.R
\name{get.cell.type.reference}
\alias{get.cell.type.reference}
\title{Get cell type reference}
\usage{
get.cell.type.reference(name)
}
\arguments{
\item{name, }{string with predefined datasets andrews and bakulski cord
blood, blood gse35069, blood gse35069 chen, blood gse35069 complete,
"combined cord blood", "cord blood gse68456", "gervin and lyle cord blood",
"guintivano dlpfc" or "saliva gse48472"}
}
\value{
name and reference.globals
}
\description{
Get cell type reference
}
\details{
ORIGINAL AUTHOR: Matthew Suderman
at githug : https://github.com/perishky/meffil
The original meffil.list.cell.type.references and 
get.cell.type.reference function from meffil v1.0.0
}
\examples{
name <- "andrews and bakulski cord blood"
get.cell.type.reference(name)

}
