% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadClocks.R
\name{load_DNAm_Clocks_data}
\alias{load_DNAm_Clocks_data}
\title{Loads DNAm clock data from methylclockData}
\usage{
load_DNAm_Clocks_data()
}
\value{
void
}
\description{
Loads DNAm clock data from methylclockData
}
\examples{
load_DNAm_Clocks_data()
}
