% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadClocksGA.R
\name{load_DNAmGA_Clocks_data}
\alias{load_DNAmGA_Clocks_data}
\title{Loads DNAmGA clock data from methylclockData}
\usage{
load_DNAmGA_Clocks_data()
}
\value{
void
}
\description{
Loads DNAmGA clock data from methylclockData
}
\examples{
load_DNAm_Clocks_data()
}
