% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffil.list.cell.type.references.R
\name{meffil.list.cell.type.references}
\alias{meffil.list.cell.type.references}
\title{List of available cell type references}
\usage{
meffil.list.cell.type.references()
}
\value{
a list with reference globals
}
\description{
List of available cell type references
}
\details{
ORIGINAL AUTHOR: Matthew Suderman
The original meffil.list.cell.type.references and get.cell.type.reference function from meffil v1.0.0
at githug : https://github.com/perishky/meffil
}
