% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffilListCellTypeReferences.R
\name{meffilListCellTypeReferences}
\alias{meffilListCellTypeReferences}
\title{List of available cell type references}
\usage{
meffilListCellTypeReferences()
}
\value{
a list with reference globals
}
\description{
List of available cell type references
}
\details{
ORIGINAL AUTHOR: Matthew Suderman
The original meffilListCellTypeReferences and 
getCellTypeReference function from meffil v1.0.0
at githug : https://github.com/perishky/meffil
}
\examples{
meffilListCellTypeReferences()

}
