% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MethylationDataExample55}
\alias{MethylationDataExample55}
\title{MethylationDataExample55}
\format{
A data frame with 27578 cpgs and 17 samples
}
\usage{
data(MethylationDataExample55)
}
\description{
A dataset containing data for vignette examples
}
\examples{
data(MethylationDataExample55)
}
\keyword{datasets}
