% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorClocks.R
\name{plotCorClocks}
\alias{plotCorClocks}
\title{Plot correlation among DNAm clockx}
\usage{
plotCorClocks(x, ...)
}
\arguments{
\item{x}{a tible or data.frame with the different DNAm clocks}

\item{...}{other arguments to be passs through function 'chart.Correlation' from 'PerformanceAnalytics' package}
}
\description{
Plot correlation among DNAm clockx
}
\details{
To be supplied
}
