% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probeAnnotation21kdatMethUsed}
\alias{probeAnnotation21kdatMethUsed}
\title{probeAnnotation21kdatMethUsed}
\format{
A data frame with 21368 CpGs and 7 variables
}
\usage{
data(probeAnnotation21kdatMethUsed)
}
\description{
A dataset containing probeAnnotation21kdatMethUsed (REMOVE??)
}
\examples{
data(probeAnnotation21kdatMethUsed)
}
\keyword{datasets}
