% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_vars}
\alias{progress_vars}
\title{PROGRESS cohort variables}
\format{
A data frame with 150 obs. and 3 variables
}
\usage{
data(progress_vars)
}
\description{
The PROGRESS cohort data is available in the additional file 8 of :
Knight, A.K., Craig, J.M., Theda, C. et al. 
An epigenetic clock for gestational age at birth based on blood
methylation data. 
Genome Biol 17, 206 (2016). 
https://doi.org/10.1186/s13059-016-1068-z
}
\details{
A dataset containing data from the PROGRESS (Programming 
Research in Obesity, Growth, Environment and Social Stressors) cohort
}
\examples{
data(progress_vars)
}
\keyword{datasets}
