% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestDataset}
\alias{TestDataset}
\title{TestDataset}
\format{
A data frame with 27236 CpGs and 4 variables
}
\usage{
data(TestDataset)
}
\description{
A dataset containing TestDataset
}
\examples{
data(TestDataset)
}
\keyword{datasets}
