% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefHorvath}
\alias{coefHorvath}
\title{Horvath clock coefficients}
\format{
A data frame with 354 CpGs and 23 variables
}
\usage{
data(coefHorvath)
}
\description{
A dataset containing the coefficients and CpGs information from Horvath clock
}
\references{
\url{https://doi.org/10.1371/journal.pone.0014821}
\describe{
  \item{CpGMarker}{CpGs id}
  \item{CoefficientTraining}{Coefficient from training }
  \item{CoefficientTrainingShrunk}
  \item{varByCpG}
  \item{minByCpG}
  \item{maxByCpG}
  \item{medianByCpG}
  \item{medianByCpGYoung}
  \item{medianByCpGOld}
  \item{Gene_ID}
  \item{GenomeBuild}
  \item{Chr}{Chromosome }
  \item{MapInfo}
  \item{SourceVersion}
  \item{TSS_Coordinate}
  \item{Gene_Strand}
  \item{Symbol}
  \item{Synonym}
  \item{Accession}
  \item{GID}
  \item{Annotation}
  \item{Product}
  \item{Marginal.Age.Relationship}
}
}
\keyword{datasets}
