% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefLeeGA}
\alias{coefLeeGA}
\title{Lee Gestational Age clock coefficients}
\format{
A data frame with 1126 CpGs and 5 variables
\describe{
  \item{Coefficient_CPC}{Coefficient_CPC}
  \item{Coefficient_RPC}{RPC Coefficient }
  \item{Coefficient_refined_RPC}{Coefficient refined RPC}
  \item{Coefficient_sex_classifier}{Coefficient sex classifier}
  \item{CpGmarker}{CpGmarker}
}
}
\usage{
data(coefLeeGA)
}
\description{
A dataset containing the coefficients and CpGs information from Lee GA clock
}
\examples{
data(coefLeeGA)
}
\references{
\url{https://dx.doi.org/10.18632\%2Faging.102049}
}
\keyword{datasets}
