% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefMayneGA}
\alias{coefMayneGA}
\title{Mayne Gestational Age clock coefficients}
\format{
A data frame with 1126 CpGs and 5 variables
\describe{
  \item{CpGMarker}{CpGs id}
  \item{CoefficientTraining}{Training Coefficient }
  \item{Gene}{Gene}
  \item{Chromosome}{Chromosome }
  \item{Position}{Position}
  \item{Correlation.with.GA}{}
}
}
\usage{
data(coefMayneGA)
}
\description{
A dataset containing the coefficients and CpGs information from Mayne GA clock
}
\references{
\url{https://dx.doi.org/10.2217\%2Fepi-2016-0103}
}
\keyword{datasets}
