% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmAge.R
\name{DNAmAge}
\alias{DNAmAge}
\title{Estimating DNAm age using Steve Horvath's method.}
\usage{
DNAmAge(x, fastImp = FALSE, cell.type = TRUE,
  cell.type.reference = "blood gse35069", age)
}
\arguments{
\item{x}{matrix ... ponemos un MethlylSet? rows -> CpGs columns 1ª nombres}

\item{fastImp}{Is fast imputation performed if necessary? (see details). Default is FALSE}

\item{cell.type}{Are IEAA and EEAA computed? Default is TRUE.}

\item{cell.type.reference}{Used when 'cell.type' is TRUE. Default is "blood gse35069". See 'meffil::meffil.list.cell.type.references()' for possible values.}

\item{age}{individual's chronological age. Required when 'cell.type' is TRUE.}
}
\description{
Estimating DNAm age using Steve Horvath's method.
}
\details{
Imputation is performed when having missing data.
         Fast imputation is performed by ...
         what about imputing only when CpGs for the clock are missing?
}
