% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmAge.R
\name{DNAmAge}
\alias{DNAmAge}
\title{DNAm age estimation using different DNA methylation clocks.}
\usage{
DNAmAge(x, GestationalAge = FALSE, toBetas = FALSE, fastImp = FALSE,
  normalize = FALSE, cell.count = TRUE,
  cell.count.reference = "blood gse35069 complete", age)
}
\arguments{
\item{x}{data.frame (Individual in columns, CpGs in rows, CpG names in first colum - i.e. Horvath's format), ExpressionSet or GenomicRatioSet.}

\item{GestationalAge}{Is gestational age clock estimated? Default is FALSE.}

\item{toBetas}{Should data be transformed to beta values? Default is FALSE. If TRUE, it implies data are M values.}

\item{fastImp}{Is fast imputation performed if necessary? (see details). Default is FALSE}

\item{normalize}{Is Horvath's normalization performed? By default is FALSE}

\item{cell.count}{Are IEAA and EEAA computed? Default is TRUE.}

\item{cell.count.reference}{Used when 'cell.count' is TRUE. Default is "blood gse35069 complete". See 'meffil::meffil.list.cell.count.references()' for possible values.}

\item{age}{individual's chronological age. Required when 'cell.count' is TRUE.}
}
\description{
DNAm age estimation using different DNA methylation clocks.
}
\details{
Imputation is performed when having missing data.
         Fast imputation is performed by ...
         what about imputing only when CpGs for the clock are missing?
}
