% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylclock.R
\docType{package}
\name{methylclock}
\alias{methylclock}
\title{methylclock}
\description{
Package to estimate DNA methylation age (DNAmAge) using different methylation clocks.
}
\author{
Juan R Gonzalez <juanr.gonzalez@isglobal.org>
}
