% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmAge.R
\name{plotDNAmAge}
\alias{plotDNAmAge}
\title{Plot DNAm age estimation vs chronological age.}
\usage{
plotDNAmAge(x, y, tit = "Horvath's method", ...)
}
\arguments{
\item{x}{DNAm age estimation}

\item{y}{Chronological age}

\item{tit}{Plot title. Default is "Horvath's method".}
}
\description{
Plot DNAm age estimation vs chronological age.
}
