% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_data}
\alias{progress_data}
\title{progress_data}
\format{
A data frame with 148 obs. and 151 variables
}
\usage{
data(progress_data)
}
\description{
A dataset containing progress_data
}
\examples{
data(progress_data)
}
\keyword{datasets}
