% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_vars}
\alias{progress_vars}
\title{progress_vars}
\format{
A data frame with 150 obs. and 3 variables
}
\usage{
data(progress_vars)
}
\description{
A dataset containing progress_vars
}
\examples{
data(progress_vars)
}
\keyword{datasets}
